# Optimize SiLU (Swish) op in PyTorch