--- a/docs/source/scripts/build_activation_images.py
+++ b/docs/source/scripts/build_activation_images.py
@@ -36,6 +36,7 @@ functions = [
     'ReLU6',
     'RReLU',
     'SELU',
+    'SiLU',
     'CELU',
     'GELU',
     'Sigmoid',